+++ src/Entity/LabInstance.php
@@ -7,10 +7,11 @@ use Doctrine\ORM\Mapping as ORM;
 use Doctrine\Common\Collections\Criteria;
 use Doctrine\Common\Collections\Collection;
 use JMS\Serializer\Annotation as Serializer;
+use Doctrine\Common\Collections\ArrayCollection;
 
 /**
  * @ORM\Entity(repositoryClass="App\Repository\LabInstanceRepository")
- * @Serializer\XmlRoot("instance")
+ * @Serializer\XmlRoot("lab_instance")
  */
 class LabInstance extends Instance
 {
@@ -25,15 +26,28 @@ class LabInstance extends Instance
 
     /**
      * @ORM\ManyToOne(targetEntity="App\Entity\Lab", inversedBy="instances")
-     * @Serializer\Groups({"lab"})
+     * @Serializer\Groups({"lab", "start_lab", "stop_lab"})
      */
     protected $lab;
 
     /**
      * @ORM\ManyToOne(targetEntity="App\Entity\User", inversedBy="labInstances")
-     * @Serializer\Groups({"user"})
+     * @Serializer\Groups({"user", "start_lab", "stop_lab"})
      */
     protected $user;
+
+    /**
+     * @ORM\OneToMany(targetEntity="App\Entity\DeviceInstance", mappedBy="labInstance")
+     * @Serializer\XmlList(inline=true, entry="device_instance")
+     * @Serializer\Groups({"lab", "start_lab", "stop_lab"})
+     */
+    protected $deviceInstances;
+
+    public function __construct()
+    {
+        parent::__construct();
+        $this->deviceInstances = new ArrayCollection();
+    }
 
     public function getId(): ?int
     {
@@ -66,45 +80,91 @@ class LabInstance extends Instance
 
     /**
      * @Serializer\VirtualProperty()
-     * @Serializer\Groups({"lab"})
+     * @Serializer\Groups({"lab", "start_lab", "stop_lab"})
      * @Serializer\XmlAttribute
      */
-    public function getUserId(): ?string
+    public function getUserId(): ?int
     {
         return $this->user->getId();
     }
 
     /**
-     * Get device instances associated to this lab
+     * Generate a bridge name with instance UUID.
+     * 
+     * @Serializer\VirtualProperty()
+     * @Serializer\Groups({"lab", "start_lab", "stop_lab"})
+     * @Serializer\XmlAttribute
+     */
+    public function getBridgeName(): string
+    {
+        return "br-" . substr($this->uuid, 0, 8);
+    }
+
+    /**
+     * Get device instance associated to this lab instance
+     *
+     * @return DeviceInstance
+     */
+    public function getDeviceInstance($device): ?DeviceInstance
+    {
+        $criteria = Criteria::create()
+            ->where(Criteria::expr()->eq("device", $device));
+
+        $deviceInstance = ($this->deviceInstances !== null) ? $this->deviceInstances->matching($criteria)->first() : null;
+        
+        return $deviceInstance ?: null;
+    }
+
+    /**
+     * Get device instances associated to this lab instance
      *
      * @return Collection|DeviceInstance[]
      */
     public function getDeviceInstances(): Collection
     {
-        // $deviceInstances = $this->lab->getDeviceInstances()->filter(function(DeviceInstance $deviceInstance) {
-        //     return $deviceInstance->getUser() == $this->user;
-        // });
+        return $this->deviceInstances;
+    }
 
+    /**
+     * Get device instance associated to this lab instance and the current user
+     *
+     * @return DeviceInstance
+     */
+    public function getUserDeviceInstance($device): ?DeviceInstance
+    {
+        $criteria = Criteria::create()
+            ->where(Criteria::expr()->eq("device", $device))
+            ->andWhere(Criteria::expr()->eq("user", $this->user))
+        ;
+
+        $deviceInstance = ($this->deviceInstances !== null) ? $this->deviceInstances->matching($criteria)->first() : null;
+        
+        return $deviceInstance ?: null;
+    }
+
+    /**
+     * Get device instances associated to this lab instance and the current user
+     *
+     * @return Collection|DeviceInstance[]
+     */
+    public function getUserDeviceInstances(): Collection
+    {
         $criteria = Criteria::create()
             ->where(Criteria::expr()->eq("user", $this->user))
         ;
 
-        $deviceInstances = $this->lab->getDeviceInstances()->matching($criteria);
+        $deviceInstances = $this->deviceInstances->matching($criteria);
 
         return $deviceInstances;
     }
 
     /**
-     * Get network interface instances associated to this lab
+     * Get network interface instances associated to this lab instance
      *
      * @return Collection|NetworkInterfaceInstance[]
      */
     public function getNetworkInterfaceInstances(): Collection
     {
-        // $deviceInstances = $this->lab->getDeviceInstances()->filter(function(DeviceInstance $deviceInstance) {
-        //     return $deviceInstance->getUser() == $this->user;
-        // });
-
         $criteria = Criteria::create()
             ->where(Criteria::expr()->eq("user", $this->user))
         ;
@@ -112,5 +172,33 @@ class LabInstance extends Instance
         $networkInterfaceInstances = $this->lab->getNetworkInterfaceInstances()->matching($criteria);
 
         return $networkInterfaceInstances;
+    }
+
+    public function addDeviceInstance(DeviceInstance $deviceInstance): self
+    {
+        if (!$this->deviceInstances->contains($deviceInstance)) {
+            $this->deviceInstances[] = $deviceInstance;
+            $deviceInstance->setLabInstance($this);
+        }
+
+        return $this;
+    }
+
+    public function removeDeviceInstance(DeviceInstance $deviceInstance): self
+    {
+        if ($this->deviceInstances->contains($deviceInstance)) {
+            $this->deviceInstances->removeElement($deviceInstance);
+            // set the owning side to null (unless already changed)
+            if ($deviceInstance->getLabInstance() === $this) {
+                $deviceInstance->setLabInstance(null);
+            }
+        }
+
+        return $this;
+    }
+
+    public function hasDeviceInstance(): bool
+    {
+        return $this->deviceInstances->count() > 0;
     }
 }