+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ### Features
 - New installer : set your environment directly from CLI installer and handle properly installation workflow.
+- A device may now be used in different labs by the same user.
 
 ### Bug fixes
 -